+++ context/app/static/js/components/workspaces/NewWorkspaceDialog/NewWorkspaceDialogFromSelections.tsx
@@ -9,14 +9,13 @@ import InputAdornment from '@mui/material/InputAdornment';
 import NewWorkspaceDialog from 'js/components/workspaces/NewWorkspaceDialog';
 import WorkspaceField from 'js/components/workspaces/WorkspaceField';
 import ErrorMessages from 'js/shared-styles/alerts/ErrorMessages';
-import useProtectedDatasetsForm from 'js/components/workspaces/NewWorkspaceDialog/useProtectedDatasetsForm';
 import { useHandleCopyClick } from 'js/hooks/useCopyText';
 import { useSelectableTableStore } from 'js/shared-styles/tables/SelectableTableProvider';
-import { useCreateWorkspaceForm } from './useCreateWorkspaceForm';
+import { useCreateWorkspaceDatasets, useCreateWorkspaceForm } from './useCreateWorkspaceForm';
 
 function NewWorkspaceDialogFromSelections() {
   const { errorMessages, protectedHubmapIds, removeProtectedDatasets, protectedRows, selectedRows } =
-    useProtectedDatasetsForm();
+    useCreateWorkspaceDatasets();
   const { deselectRows } = useSelectableTableStore();
 
   const { control, errors, setDialogIsOpen, ...rest } = useCreateWorkspaceForm({});
